🚚(frontend) move CourseProductItem into SyllabusCourseRunsList
Since we ported course run list in React, `CourseProductItem` is no more a
widget but a component used only in the scope of `SyllabusCourseRunsList` widget
. So we move `CourseProductItem` into `SyllabusCourseRunsList` components folder
 then remove its import from the frontend application index module.
