+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/_styles.scss
