+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PaymentButton/_styles.scss
