+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PaymentButton/index.spec.tsx
