+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PaymentInterfaces/__mocks__/index.tsx
