+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PaymentInterfaces/index.spec.tsx
