+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PurchaseButton/index.spec.tsx
