+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/RegisteredCreditCard/_styles.scss
