+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/RegisteredCreditCard/index.spec.tsx
