+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnel/_styles.scss
