+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepPayment/index.spec.tsx
