+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepResume/_styles.scss
