+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepValidation/CourseRunsList.tsx
