+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/StepBreadcrumb/_styles.scss
