+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/StepBreadcrumb/index.spec.tsx
