+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/contexts/CourseProductContext/index.spec.tsx
