+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/hooks/useStepManager/index.ts
