+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/index.spec.tsx
