+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/index.tsx
