+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/CourseProductItem/types/payments/payplug.d.ts
