+++ src/frontend/scss/components/_index.scss
@@ -11,16 +11,6 @@
 @import '../../js/pages/DashboardPreferences/styles';
 @import '../../js/pages/TeacherCourseDashboardLoader/CourseRunList/styles';
 @import '../../js/pages/TeacherTrainingDashboard/styles.scss';
-@import '../../js/widgets/CourseProductItem/components/CourseProductCertificateItem/styles';
-@import '../../js/widgets/CourseProductItem/components/CourseProductCourseRuns/styles';
-@import '../../js/widgets/CourseProductItem/components/PaymentButton/styles';
-@import '../../js/widgets/CourseProductItem/components/RegisteredCreditCard/styles';
-@import '../../js/widgets/CourseProductItem/components/SaleTunnel/styles';
-@import '../../js/widgets/CourseProductItem/components/SaleTunnelStepPayment/styles';
-@import '../../js/widgets/CourseProductItem/components/SaleTunnelStepResume/styles';
-@import '../../js/widgets/CourseProductItem/components/SaleTunnelStepValidation/styles';
-@import '../../js/widgets/CourseProductItem/components/StepBreadcrumb/styles';
-@import '../../js/widgets/CourseProductItem/styles';
 @import '../../js/widgets/Dashboard/styles';
 @import '../../js/widgets/Dashboard/components/DashboardAvatar/styles';
 @import '../../js/widgets/Dashboard/components/DashboardBox/styles';
@@ -36,6 +26,16 @@
 @import '../../js/widgets/Search/components/SearchFilterGroup/styles';
 @import '../../js/widgets/Search/components/SearchFilterGroupModal/styles';
 @import '../../js/widgets/Search/components/SearchFiltersPane/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/CourseProductCertificateItem/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/CourseProductCourseRuns/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/PaymentButton/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/RegisteredCreditCard/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnel/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepPayment/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepResume/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/SaleTunnelStepValidation/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/components/StepBreadcrumb/styles';
+@import '../../js/widgets/SyllabusCourseRunsList/components/CourseProductItem/styles';
 @import '../../js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/CourseRunUnenrollmentButton/styles';
 @import '../../js/widgets/SyllabusCourseRunsList/components/CourseRunEnrollment/styles';
 @import '../../js/widgets/SyllabusCourseRunsList/components/CourseWishButton/styles';
